moved main() to own file, jupyter-specific code in own file
