+++ src/python/impactx/dashboard/start.py
@@ -1,0 +1,16 @@
+from trame.app import get_server
+
+# -----------------------------------------------------------------------------
+# Trame setup
+# -----------------------------------------------------------------------------
+
+server = get_server(client_type="vue2")
+state, ctrl = server.state, server.controller
+
+# -----------------------------------------------------------------------------
+# Main
+# -----------------------------------------------------------------------------
+
+def main():
+    server.start()
+    return 0
